# docs/swagger.yaml
basePath: /
definitions:
  model.Comment:
    properties:
      commentID:
        type: string
      createdAt:
        type: string
      message:
        type: string
      photoID:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      code:
        type: integer
      errors: {}
      status:
        type: string
    type: object
...
  model.PhotoDeleteRes:
    properties:
      photo_id:
        type: string
    type: object
  model.PhotoResponse:
    properties:
      caption:
        type: string
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      photo_id:
        type: string
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.PhotoUpdateReq:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    required:
    - photo_url
    - title
    type: object
  model.PhotoUpdateRes:
    properties:
      photo_id:
        type: string
    type: object
...
  model.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a MyGram API.
  title: MyGram API
  version: "1.0"
paths:
...
  /photos/{photo_id}:
    delete:
      consumes:
      - application/json
      description: delete photo
      parameters:
      - description: PhotoID
        in: path
        name: photo_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PhotoDeleteRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: delete photo
      tags:
      - Photo
    get:
      consumes:
      - application/json
      description: get one photo
      parameters:
      - description: PhotoID
        in: path
        name: photo_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PhotoResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: get one photo
      tags:
      - Photo
    put:
      consumes:
      - application/json
      description: update photo
      parameters:
      - description: PhotoID
        in: path
        name: photo_id
        required: true
        type: string
      - description: request is required
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PhotoUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PhotoUpdateRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/model.ErrorResponse'
            - properties:
                errors:
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: update photo
      tags:
      - Photo
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"